Implemented constants in Functions
This is an idea from Lua. Functions now have a vector of Values which are the constant values used in that function. So now, rather than having a PUSHS with a parameter that is the StringLiteral id, there is a PUSHK with a single byte ConstantId parameter. The actual StringLiteral is stored in a Value in the constants vector, and the ConstantId is an index into that array. This will allow fewer opcodes and more efficiency and will allow me to implement a register based virtual machine

+++ m8rscript/Parser.h
@@ -109,7 +109,7 @@ public:
     void programEnd();
         
     void emit(StringLiteral::Raw value);
-    void emit(uint64_t value);
+    void emit(uint32_t value);
     void emit(Float value);
     void emit(Op value);
